requirements.txt
opencv-python==4.9.0.80
mediapipe==0.10.9
numpy==1.26.4
google-generativeai==NO IDEA JUST PIP INSTALL IT
